+++ model_navigator/api/utilities.py
@@ -14,7 +14,6 @@
 # noqa: D104
 """Public utilities for the Model Navigator API."""
 
-import functools
 import json
 import logging
 import pathlib
@@ -112,9 +111,8 @@ def find_max_batch_size_till_oom(
         runner = get_runner(runner_name)(
             model=model,
             input_metadata=input_metadata,
-            output_metadata=TensorMetadata(),
+            output_metadata=None,
         )  # pytype: disable=not-instantiable
-        runner.infer = functools.partial(runner.infer, return_raw_outputs=True, check_inputs=False)
         try:
             LOGGER.info("Starting max batch size search.")
             MaxBatchSizeFinder(