+++ tests/functional/L0_Optimize_and_Save_jax_linear/test.sh
@@ -19,6 +19,7 @@ THIS_SCRIPT_PATH="$(realpath --relative-to="$(pwd)" "$0")"
 TEST_MODULE="$(dirname "${THIS_SCRIPT_PATH}"|sed 's/\//./g').test"
 
 pip install --upgrade flax "jax[cuda12_pip]<0.4.16" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
+pip install "ml-dtypes==0.2.0"
 
 export XLA_PYTHON_CLIENT_PREALLOCATE=false
 python -m"${TEST_MODULE}" \