+++ tests/functional/L0_Optimize_and_Save_tensorrt_identity/test.sh
@@ -18,9 +18,6 @@ set -ex
 THIS_SCRIPT_PATH="$(realpath --relative-to="$(pwd)" "$0")"
 TEST_MODULE="$(dirname "${THIS_SCRIPT_PATH}"|sed 's/\//./g').test"
 
-pip install --upgrade torch transformers flax "jax[cuda12_pip]<0.4.16" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
-pip install "ml-dtypes==0.2.0"
-
 python -m"${TEST_MODULE}" \
     --status $(pwd)/status.yaml \
     --verbose